Move filesystem to mock namespace
