+++ tests/units/classes/mock/filesystem/file.php
@@ -1,49 +1,73 @@
 <?php
-namespace mageekguy\atoum\tests\units\filesystem;
+namespace mageekguy\atoum\tests\units\mock\filesystem;
 
 use
 	mageekguy\atoum,
-	mageekguy\atoum\filesystem\node as testedClass
+	mageekguy\atoum\mock\filesystem\file as testedClass
 ;
 
 require_once __DIR__ . '/../../runner.php';
 
 use mageekguy\atoum\mock\stream;
 
-class node extends atoum\test
+class file extends atoum\test
 {
 	public function test__construct()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node())
+			->testedClass
+				->isSubClassOf('\\mageekguy\\atoum\\mock\\filesystem\\node')
+			->if($object = new testedClass())
 			->then
 				->string($object->getName())->isNotEmpty()
 				->variable($object->getParent())->isNull()
 				->object($object->getStream())->isInstanceOf('\\mageekguy\\atoum\\mock\\stream\\controller')
-			->if($object = new \mock\mageekguy\atoum\filesystem\node($name = uniqid()))
+				->boolean(is_file($object))->isTrue()
+				->string(file_get_contents($object))->isEmpty()
+			->if($object = new testedClass($name = uniqid()))
 			->then
 				->string($object->getName())->isEqualTo($name)
 				->variable($object->getParent())->isNull()
 				->object($object->getStream())->isInstanceOf('\\mageekguy\\atoum\\mock\\stream\\controller')
+				->boolean(is_file($object))->isTrue()
+				->string(file_get_contents($object))->isEmpty()
 			->if($this->mockGenerator->shunt('__construct'))
-			->and($node = new \mock\mageekguy\atoum\filesystem\node())
+			->and($node = new \mock\mageekguy\atoum\mock\filesystem\node())
 			->and($node->getMockController()->getStream = stream::get())
-			->and($object = new \mock\mageekguy\atoum\filesystem\node($name, $node))
+			->and($object = new testedClass($name, $node))
 			->then
 				->string($object->getName())->isIdenticalTo($name)
 				->object($object->getParent())->isIdenticalTo($node)
 				->object($object->getStream())->isInstanceOf('\\mageekguy\\atoum\\mock\\stream\\controller')
-			->if($object = new \mock\mageekguy\atoum\filesystem\node($name, $node))
+				->boolean(is_file($object))->isTrue()
+				->string(file_get_contents($object))->isEmpty()
+			->if($object = new testedClass($name, $node))
 			->then
 				->string($object->getName())->isIdenticalTo($name)
 				->object($object->getParent())->isIdenticalTo($node)
+				->boolean(is_file($object))->isTrue()
+				->string(file_get_contents($object))->isEmpty()
+		;
+	}
+
+	public function testGetContent()
+	{
+		$this
+			->if($object = new testedClass())
+			->then
+				->object($object->setContent($content = uniqid()))->isIdenticalTo($object)
+				->string(file_get_contents($object))->isEqualTo($content)
+			->if($object = new testedClass())
+			->then
+				->object($object->content($content = uniqid()))->isIdenticalTo($object)
+				->string(file_get_contents($object))->isEqualTo($content)
 		;
 	}
 
 	public function testGetName()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node($name = uniqid()))
+			->if($object = new testedClass($name = uniqid()))
 			->then
 				->string($object->getName())->isEqualTo($name)
 		;
@@ -52,7 +76,7 @@ class node extends atoum\test
 	public function testGetStream()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node())
+			->if($object = new testedClass())
 			->then
 				->object($object->getStream())->isInstanceOf('\\mageekguy\\atoum\\mock\\stream\\controller')
 		;
@@ -61,13 +85,13 @@ class node extends atoum\test
 	public function testGetParent()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node())
+			->if($object = new testedClass())
 			->then
 				->variable($object->getParent())->isNull()
 			->if($this->mockGenerator->shunt('__construct'))
-			->and($node = new \mock\mageekguy\atoum\filesystem\node())
+			->and($node = new \mock\mageekguy\atoum\mock\filesystem\node())
 			->and($node->getMockController()->getStream = stream::get())
-			->and($object = new \mock\mageekguy\atoum\filesystem\node(uniqid(), $node))
+			->and($object = new testedClass(uniqid(), $node))
 			->then
 				->object($object->getParent())->isIdenticalTo($node)
 		;
@@ -76,22 +100,26 @@ class node extends atoum\test
 	public function testEnd()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node())
+			->if($object = new testedClass())
 			->then
 				->variable($object->end())->isNull()
+				->variable($object->close())->isNull()
+				->variable($object->parent())->isNull()
 			->if($this->mockGenerator->shunt('__construct'))
-			->and($node = new \mock\mageekguy\atoum\filesystem\node())
+			->and($node = new \mock\mageekguy\atoum\mock\filesystem\node())
 			->and($node->getMockController()->getStream = stream::get())
-			->and($object = new \mock\mageekguy\atoum\filesystem\node(uniqid(), $node))
+			->and($object = new testedClass(uniqid(), $node))
 			->then
 				->object($object->end())->isIdenticalTo($node)
+				->object($object->close())->isIdenticalTo($node)
+				->object($object->parent())->isIdenticalTo($node)
 		;
 	}
 
 	public function testReferencedBy()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node())
+			->if($object = new testedClass())
 			->and($reference = null)
 			->then
 				->object($object->referencedBy($reference))->isIdenticalTo($object)
@@ -102,8 +130,8 @@ class node extends atoum\test
 	public function test__call()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node(uniqid()))
-			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\stream\controller())
+			->if($object = new \mock\mageekguy\atoum\mock\filesystem\file($name = uniqid()))
+			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\mock\stream\controller($name))
 			->and($stream->getMockController()->invoke = function() {})
 			->then
 				->variable($object->foo())
@@ -118,8 +146,8 @@ class node extends atoum\test
 	public function test__get()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node(uniqid()))
-			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\stream\controller())
+			->if($object = new \mock\mageekguy\atoum\mock\filesystem\file($name = uniqid()))
+			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\mock\stream\controller($name))
 			->and($stream->getMockController()->__get = function() {})
 			->then
 				->variable($object->foo)
@@ -131,8 +159,8 @@ class node extends atoum\test
 	public function test__set()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node(uniqid()))
-			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\stream\controller())
+			->if($object = new \mock\mageekguy\atoum\mock\filesystem\file($name = uniqid()))
+			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\mock\stream\controller($name))
 			->and($stream->getMockController()->__set = function() {})
 			->then
 				->variable($object->foo = $arg = uniqid())
@@ -144,9 +172,9 @@ class node extends atoum\test
 	public function test__isset()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node(uniqid()))
-			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\stream\controller())
-			->and($stream->getMockController()->__set = function() {})
+			->if($object = new \mock\mageekguy\atoum\mock\filesystem\file($name = uniqid()))
+			->and($object->getMockController()->getStream = $stream = new \mock\mageekguy\atoum\mock\stream\controller($name))
+			->and($stream->getMockController()->__isset = function() {})
 			->then
 				->boolean(isset($object->foo))->isFalse()
 				->mock($stream)
@@ -157,7 +185,7 @@ class node extends atoum\test
 	public function test__toString()
 	{
 		$this
-			->if($object = new \mock\mageekguy\atoum\filesystem\node($name = uniqid()))
+			->if($object = new testedClass($name = uniqid()))
 			->then
 				->castToString($object)->isEqualTo('atoum://' . $name)
 		;
